+++ zebra/zebra_router.c
@@ -29,9 +29,12 @@
 #include "zebra_pbr.h"
 #include "zebra_vxlan.h"
 #include "zebra_mlag.h"
+#include "zebra_nhg.h"
+#include "debug.h"
 
 struct zebra_router zrouter = {
 	.multipath_num = MULTIPATH_NUM,
+	.ipv4_multicast_mode = MCAST_NO_CONFIG,
 };
 
 static inline int
@@ -185,6 +188,19 @@ uint32_t zebra_router_get_next_sequence(void)
 	return 1
 	       + atomic_fetch_add_explicit(&zrouter.sequence_num, 1,
 					   memory_order_relaxed);
+}
+
+void multicast_mode_ipv4_set(enum multicast_mode mode)
+{
+	if (IS_ZEBRA_DEBUG_RIB)
+		zlog_debug("%s: multicast lookup mode set (%d)", __func__,
+			   mode);
+	zrouter.ipv4_multicast_mode = mode;
+}
+
+enum multicast_mode multicast_mode_ipv4_get(void)
+{
+	return zrouter.ipv4_multicast_mode;
 }
 
 void zebra_router_terminate(void)
